Delete

+++ Back/src/main/java/ar/edu/utn/frc/tup/lciii/dtos/SearchResponce.java
@@ -10,6 +10,6 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class SearchResponce {
-    Long countTotal;
+    int countTotal;
     List<PublicationMinDto> list;
 }
